#208 Server - Model: Song Section - Default Confidence to Uint and 0

+++ repertoire.server/domain/usecase/song/create_song.go
@@ -87,7 +87,7 @@ func (c CreateSong) createSections(request []requests.CreateSectionRequest, song
 		sections = append(sections, model.SongSection{
 			ID:                uuid.New(),
 			Name:              sectionRequest.Name,
-			Confidence:        uint(model.DefaultSongSectionConfidence),
+			Confidence:        model.DefaultSongSectionConfidence,
 			SongSectionTypeID: sectionRequest.TypeID,
 			Order:             uint(i),
 			SongID:            songID,
